openfoldpp/requirements.txt
# OpenFold++ Requirements

# Core ML/Scientific Computing
torch>=2.0.0
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Protein/Bioinformatics
biopython>=1.79
biotite>=0.37.0

# Molecular Dynamics (optional)
openmm>=7.7.0
mdtraj>=1.9.7

# Language Models (optional)
transformers>=4.20.0
tokenizers>=0.13.0
fair-esm>=2.0.0  # Facebook ESM-2 protein language model
sentencepiece>=0.1.97  # For ESM tokenization

# Performance Optimizations (optional)
flash-attn>=2.0.0
bitsandbytes>=0.39.0
performer-pytorch>=1.1.4

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
requests>=2.28.0
pathlib2>=2.3.7

# Development/Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# CUDA Kernels (optional)
ninja>=1.10.0
